features: no on-device editing; yes broader device support
Customizing the watch face on-device is introduced in Connect IQ API
version 5.1, which is way too new for some of the devices we want to
target (e.g. the Vivoactive 3 is v3.1, and the Vivoactive HR is v2.4).

Instead, we want to aim for broader device support, though different
screen shapes could be weird. For the moment, that's a later™ problem.