core,layout,drawables: properly unload images, draw layouts manually
Bitmaps that live in layouts can't be properly unloaded, so instead we
track and draw them manually as `Graphics.BitmapType` objects. For
backgrounds, though, this means we have to draw the *entire* layout
manually — using `View.onUpdate()` doesn't work because it starts by
clearing the entire screen (wiping our background in the process).